move flee to basic needs state
