--- Assets/Scripts/Entities/EnemyDetector.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f3aa5ae8c5d5470c88c7e493b6f52b1c
-timeCreated: 1586451810
