+++ Assets/Scripts/FSM/BasicNeedState/Flee.cs
@@ -1,19 +1,17 @@
 ﻿#region
 
 using Brains;
-using Entities;
 using UnityEngine;
 using UnityEngine.AI;
 
 #endregion
 
-namespace FSM.GathererState
+namespace FSM.BasicNeedState
 {
     public class Flee : IState
     {
-        private readonly Gatherer _gatherer;
+        private readonly BasicNeeds _entity; // the person or the creature
         private NavMeshAgent _navMeshAgent;
-        private readonly EnemyDetector _enemyDetector;
         private Animator _animator;
         private readonly ParticleSystem _particleSystem;
 
@@ -21,20 +19,18 @@ namespace FSM.GathererState
         private const float FLEE_SPEED = 6F;
         private const float FLEE_DISTANCE = 5F;
 
-        public Flee(Gatherer gatherer, NavMeshAgent navMeshAgent, EnemyDetector enemyDetector, Animator animator, ParticleSystem particleSystem)
+        public Flee(BasicNeeds entity, NavMeshAgent navMeshAgent, Animator animator, ParticleSystem particleSystem)
         {
-            _gatherer = gatherer;
+            _entity = entity;
             _navMeshAgent = navMeshAgent;
-            _enemyDetector = enemyDetector;
             _animator = animator;
             _particleSystem = particleSystem;
         }
 
         public void OnEnter()
         {
-            Debug.LogFormat("{0} is fleeing", _gatherer.gameObject.name);
+            Debug.LogFormat("{0} is fleeing", _entity.gameObject.name);
             _navMeshAgent.enabled = true;
-            _gatherer.DropAllResources();
             _animator.SetBool(Constants.IsRunningHash, true);
             _initialSpeed = _navMeshAgent.speed;
             _navMeshAgent.speed = FLEE_SPEED;
@@ -52,16 +48,18 @@ namespace FSM.GathererState
 
         private Vector3 GetRandomPoint()
         {
-            var directionFromBeast = _gatherer.transform.position - _enemyDetector.GetNearestBeastPosition();
+            // TODO: get position of the enemy to avoid
+            var enemyPosition = Vector3.zero;
+
+            var directionFromBeast = _entity.transform.position - enemyPosition;
             directionFromBeast.Normalize();
 
-            var endPoint = _gatherer.transform.position + (directionFromBeast * FLEE_DISTANCE);
+            var endPoint = _entity.transform.position + (directionFromBeast * FLEE_DISTANCE);
             if (NavMesh.SamplePosition(endPoint, out var hit, 10f, NavMesh.AllAreas))
             {
                 return hit.position;
             }
-
-            return _gatherer.transform.position;
+            return _entity.transform.position;
         }
 
         public void OnExit()
